refactor(core): split View and Region modules
