+++ evidence/state-data-regression-ledger.md
@@ -12,7 +12,7 @@ evidence by itself.
 | `mixins/state.js` and owner constructors | Supplied state is the exact borrowed source; `createState()` results are owned; `{ state: undefined }` selects the lazy default | Preserved by replacement coverage | `test/unit/state-owner.spec.js` covers exact identity, borrowing, lazy defaults, a single option-accessor read, and owned disposal. Keep `getState()` identity-preserving. |
 | `mixins/state.js` and owner teardown | State event subscriptions are released before owned disposal, exactly once, even after public `off()` | Preserved by replacement coverage | `test/unit/state-owner.spec.js` and `test/unit/owned-cleanup.spec.js` assert ordering, idempotence, and that public event cleanup cannot disable owned cleanup. |
 | `modules/object.js` and `modules/application.js` | Construction failure rolls back Radio, state, Region, and listeners while preserving the construction error | Preserved by replacement coverage | `test/unit/owned-cleanup.spec.js` exercises `_initRadio`, `_initState`, `initialize`, and `stateEvents` failures with owned resources. |
-| `modules/view-region.js` (`View`), `modules/collection-view.js`, and `modules/behavior.js` | Failure after state or entity subscription rolls back all completed subscriptions and owned state; Behavior composition is removed | Preserved by replacement coverage | `test/unit/state-owner.spec.js`, `test/unit/owned-cleanup.spec.js`, Behavior rollback tests, and the failed-CollectionView-construction regression cover later DataApi and behavior-construction failures through public construction. Collection observation is released before child cleanup can mutate the source. |
+| `modules/view.js`, `modules/collection-view.js`, and `modules/behavior.js` | Failure after state or entity subscription rolls back all completed subscriptions and owned state; Behavior composition is removed | Preserved by replacement coverage | `test/unit/state-owner.spec.js`, `test/unit/owned-cleanup.spec.js`, Behavior rollback tests, and the failed-CollectionView-construction regression cover later DataApi and behavior-construction failures through public construction. Collection observation is released before child cleanup can mutate the source. |
 | `utils/subscribe-bindings.js` and `mixins/delegate-entity-events.js` | Partial event-map registration is rolled back; cleanup functions are required and idempotent | Preserved by replacement coverage | `test/unit/utils/subscribe-bindings.spec.js`, `test/unit/data-api-integration.spec.js`, and state-owner failure cases cover partial subscription failure and `MN0038`. |
 | Deleted `modules/state.js` | Function defaults, subclassing, own-property semantics, collision-safe keys and JSON, ordered notifications, nested writes, reset/unset, and terminal cleanup | Intentionally moved to `@marionette/data`, with replacement coverage | `test/unit/data-package/model.spec.js` covers function defaults, subclassing, own `undefined`, prototype-collision-safe shallow snapshots, no-op and silent writes, exact ordered and nested notifications, reset/unset/clear, and post-destroy write no-ops. Its `accepts whitespace keys as ordinary own properties` regression covers the rejection intentionally retired with MN0034. |
 | `@marionette/data` Collection | Ordered access and mutation, normalized non-silent records, silent structural changes, direct Model destruction, identity ownership, and terminal cleanup | New optional package contract with focused coverage | `packages/data/readme.md` defines the public mutation boundary. `test/unit/data-package/collection.spec.js` covers normalized records, all silent structural methods, always-emitted destroy events, multi-owner Model destruction, rollback, and post-destroy no-ops. |
